login and session management

+++ Controllers/MessageController.cs
@@ -1,5 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using System.Collections.Generic;
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
 using System;
 
 using Communicator.Services;
@@ -21,32 +21,54 @@ namespace Communicator.Controllers
 		[HttpPost]
 		public IActionResult Add(MessageRequest request)
 		{
+			if (HttpContext.Session.GetInt32("active") != 1)
+			{
+				return StatusCode(440);
+			}
 			return _service.Add(request) ? Ok() : BadRequest();
 		}
 
 		[HttpDelete]
 		public IActionResult Delete(int id)
 		{
+			if (HttpContext.Session.GetInt32("active") != 1)
+			{
+				return StatusCode(440);
+			}
 			return _service.Delete(id) ? Ok() : BadRequest();
 		}
 
 		[HttpPut]
 		public IActionResult Update(int id)
 		{
+			if (HttpContext.Session.GetInt32("active") != 1)
+			{
+				return StatusCode(440);
+			}
 			return _service.Update(id) ? Ok() : BadRequest();
 		}
 
-		[HttpGet][Route("GetByID")]
+		[HttpGet]
+		[Route("GetByID")]
 		public IActionResult GetByID(int id, bool sender)
 		{
+			if (HttpContext.Session.GetInt32("active") != 1)
+			{
+				return StatusCode(440);
+			}
 			var response = _service.GetMessage(id, sender);
 			return response != null ? Ok(response) : BadRequest();
 		}
 
-		[HttpGet][Route("GetMessages")]
-		public List<MessageResponse> GetMessages(DateTime time, int userId, int friendId)
+		[HttpGet]
+		[Route("GetMessages")]
+		public IActionResult GetMessages(DateTime time, int userId, int friendId)
 		{
-			return _service.GetMessages(time, userId, friendId);
+			if (HttpContext.Session.GetInt32("active") != 1)
+			{
+				return StatusCode(440);
+			}
+			return Ok(_service.GetMessages(time, userId, friendId));
 		}
 	}
 }
